fix dup proposal types
