+++ src/pages/[...slug].tsx
@@ -3,8 +3,8 @@ import { useRouter } from 'next/router';
 import { useMemo } from 'react';
 import { Switcher } from 'src/components/switcher/Switcher';
 import { Governance } from 'src/features/governance/components/Governance';
-import { Proposal } from 'src/features/governance/data/Proposal';
 import { getProposals } from 'src/features/governance/data/getProposals';
+import { Proposal } from 'src/features/governance/data/Proposal';
 import { Swap } from 'src/features/swap/components/Swap';
 import { Validators } from 'src/features/validators/components/List';
 import fetchValidGroups, { ValidatorGroup } from 'src/features/validators/data/fetchValidGroups';
